update help files to latest
Not sure what happened but the ./internal/help files seem to be fairly
old compared to `development` in the old qb64 repository. This updates
them all to match that repository, let me know if this looks sane. I
made sure to keep the small changes already made to them in this
repository.