Restructure source, add resource folder
Restructured the source so syntax checking and auto-format can be used whenever working with any of the source files individually.

Added a resource folder with a few default resource files, so users will always have a few static links to fall back upon when sharing code samples and demos with others.

Updated the changelog with changes we've made so far.