Add back default resource files, update readme,  correct wiki pointing back at old site
undoing part of the undo which I did with the last undo...  Gah!!

I hate github, and I think it feels the same about me!

Anyways, I think this corrects everything that was reverted out and shouldn't have been, as well as (once again) pointing help to the proper place.