Update ./internal/source when building on Windows
On Windows we don't distribute the ./setup_win.bat, so the contents of
./internal/source are not actually relevant for them. Since the QB64
symbol file will be located in it though, it is important that they get
an updated copy. It does also just make basic sense that if we're
distributing an ./internal/source with the Windows version it should
match whatever QB64 it was shipped with.