Don't include the directory in symbol file name
The symbol file was intended to just make the executable name with
".sym" and be located in ./internal/temp. Unfortunately I forgot to
remove the directory from the EXE name, so it ended up producing an
incorrect filename for the ".sym" file.

Make already provide a `notdir` function, so we can use that to get just
the executable name.