Replaced OpenAL-soft with miniaudio
This is initial change for QBPE audio. Lot of audio API is currently just stubs and have to be implemented. Basic stuff is working and we are able to build QBPE. All missing audio features will be added slowly and we'll try to achieve feature parity with the OpenAL-soft builds.