Initial _MEMSOUND implementation
This obviously needs to be tested. With this we have reached a feature-complete status (when comparing with the original OpenAL code). We gained a FLAC decoder free and Opus support will be added when miniaudio gets native support for it. I will now move on to writing format plugins for miniaudio (MIDI, MOD, XM, S3M, IT & RAD.